+++ src/app/GeneratorTeamsAppOptions.ts
@@ -15,6 +15,7 @@ export class GeneratorTeamsAppOptions {
     shouldUseSubDir: boolean;
     libraryName: string;
     'skip-install': string;
+    'no-telemetry': string;
     namespace: string;
     developer: string;
     privacy: string;