+++ src/bot/templates/src/app/{botName}.ts
@@ -1,6 +1,6 @@
 import * as builder from 'botbuilder';
 import * as teamBuilder from 'botbuilder-teams';
-import { BotDeclaration, IBot } from 'express-msteams-host';
+import { BotDeclaration, TeamsBot } from 'express-msteams-host';
 import * as debug from "debug";
 
 // Initialize debug logging module
@@ -13,7 +13,7 @@ const log = debug("msteams");
     '/api/messages',
     process.env.MICROSOFT_APP_ID,
     process.env.MICROSOFT_APP_PASSWORD)
-export class <%= botName %> implements IBot {
+export class <%= botName %> extends TeamsBot {
     public readonly Connector: teamBuilder.TeamsChatConnector;
     private readonly universalBot: builder.UniversalBot;
     private inMemoryStorage: builder.IBotStorage;
@@ -23,6 +23,7 @@ export class <%= botName %> implements IBot {
      * @param connector 
      */
     public constructor(connector: teamBuilder.TeamsChatConnector) {
+        super(connector);
         this.Connector = connector;
         this.inMemoryStorage = new builder.MemoryBotStorage();
         this.universalBot = new builder.UniversalBot(this.Connector).
@@ -43,62 +44,9 @@ export class <%= botName %> implements IBot {
         // Control messages
         this.universalBot.on('conversationUpdate', this.convUpdateHandler);
 
-        <% if (messageExtensionType == 'new' || messageExtensionType == 'existing') { %>
-            // Message Extension
-            this.Connector.onQuery('<%=messageExtensionName%>',
-                (event: builder.IEvent, query: teamBuilder.ComposeExtensionQuery, callback: (err: Error, result: teamBuilder.IComposeExtensionResponse, statusCode: number) => void) => {
-                    if (query.parameters && query.parameters[0] && query.parameters[0].name === 'initialRun') {
-                        // implement an MRU, kind of thing
-                        let firstResponse = teamBuilder.ComposeExtensionResponse.result('list').attachments([
-                            new builder.ThumbnailCard()
-                                .title('Test')
-                                .text('Test')
-                                .images([new builder.CardImage().url('<%=host%>/assets/icon.png')])
-                                .toAttachment()
-                        ]).toResponse();
-                        callback(<any>null, firstResponse, 200);
-                    }
-                    else {
-                        // Return result response
 
-                        let response = teamBuilder.ComposeExtensionResponse.result('list').attachments([
-                            new builder.ThumbnailCard()
-                                .title(`Test`)
-                                .text('test')
-                                .images([new builder.CardImage().url('<%=host%>/assets/icon.png')])
-                                .toAttachment()
-                        ]).toResponse();
-                        callback(<any>null, response, 200);
-                    }
-                });
-            // this is used when canUpdateConfiguration is set to true 
-            this.Connector.onQuerySettingsUrl(
-                (event: builder.IEvent, query: teamBuilder.ComposeExtensionQuery, callback: (err: Error, result: teamBuilder.IComposeExtensionResponse, statusCode: number) => void) => {
-                    callback(<any>null, {
-                        composeExtension: {
-                            type: "config",
-                            suggestedActions: {
-                                actions: [
-                                    {
-                                        type: "openApp",
-                                        title: "<%=messageExtensionTitle%> Configuration",
-                                        value: '<%= host %>/<%= messageExtensionName %>Config.html'
-                                    }
-                                ]
-                            }
-                        }
-                    }, 200);
-                }
-            )
-            this.Connector.onSettingsUpdate(
-                (event: builder.IEvent, query: teamBuilder.ComposeExtensionQuery, callback: (err: Error, result: teamBuilder.IComposeExtensionResponse, statusCode: number) => void) => {
-                    // take care of the setting returned from the dialog, with the value stored in state
-                    const setting = query.state;
-                    callback(<any>null, <any>null, 200);
-                }
-            )
-                <% } %>
-
+        // this row must follow the instantiation of the Message Extension object
+        super.registerMessageExtensions();
    }
 
     /**